Single quadrant scheme + Testing + Trial using normalized point
Reverse scheme - Poses in 4 sub-images emulated as if happening in a single image -- basically the format as if the result of video_pose.py cropped isn't modified.

Non-overlapping dataset turned out to have no major performance difference compared to the non-overlapped ones, and can be said to be even worse.

The current best scenario:
- Using Reverse scheme
- Direct data, overlapped, no further preprocessing.

Status:
- "Idle" is less likely to be identified compared to NormalizedOnce.
- Directional recognition is the most accurate, tho still can be considered bad.
- Specific gesture recognition might be better, but still bad.