Major Bug, Superb Performance
Basically changing from >>skel = skel * zero<< to >>skel *= zero<<, as it turns out the former is NOT saving it to the variable (meanwhile Utilities programs won't have such problem coz they're not in a function).

Now the system is just sooo good. Not even looking like a wild guess. Giving:
- 70%+ gesture group accuracy
- 80%+ security level accuracy